requirements.txt
asttokens==2.2.1
azure-ai-language-questionanswering==1.1.0
azure-ai-textanalytics==5.2.1
azure-common==1.1.28
azure-core==1.26.3
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
msrest==0.7.1
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pywin32==305
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
